--- src/main/java/DeleteCommand.java
@@ -1,11 +1,0 @@
-public class DeleteCommand extends Command {
-    Integer index;
-    public DeleteCommand(Integer index) {
-        this.index = index;
-    }
-    @Override
-    public void execute(TaskList tasks, Ui ui, Storage storage) throws DukeException {
-        ui.displayMessage(tasks.delete(index));
-        storage.save(tasks);
-    }
-}
