--- src/main/java/DocRestCommand.java
@@ -1,6 +1,0 @@
-public class DocRestCommand extends Command {
-    @Override
-    public void execute(TaskList tasks, Ui ui, Storage storage) throws DukeException {
-        ui.displayMessage(tasks.docRest());
-    }
-}
