--- src/main/java/ExitCommand.java
@@ -1,13 +1,0 @@
-public class ExitCommand extends Command {
-
-    @Override
-    public void execute(TaskList tasks, Ui ui, Storage storage) {
-        return;
-    }
-
-    @Override
-    public boolean isExit() {
-        return true;
-    }
-
-}
