+++ src/main/java/duke/command/AddCommand.java
@@ -1,0 +1,19 @@
+package duke.command;
+
+import duke.DukeException;
+import duke.Storage;
+import duke.TaskList;
+import duke.Ui;
+import duke.task.Task;
+
+public class AddCommand extends Command {
+    private Task taskToAdd;
+    public AddCommand(Task task) {
+        this.taskToAdd = task;
+    }
+    @Override
+    public void execute(TaskList tasks, Ui ui, Storage storage) throws DukeException {
+        ui.displayMessage(tasks.add(taskToAdd));
+        storage.save(tasks);
+    }
+}
