+++ src/main/java/duke/command/DocRestCommand.java
@@ -1,0 +1,13 @@
+package duke.command;
+
+import duke.DukeException;
+import duke.Storage;
+import duke.TaskList;
+import duke.Ui;
+
+public class DocRestCommand extends Command {
+    @Override
+    public void execute(TaskList tasks, Ui ui, Storage storage) throws DukeException {
+        ui.displayMessage(tasks.docRest());
+    }
+}
