+++ src/main/java/duke/command/ExitCommand.java
@@ -1,0 +1,19 @@
+package duke.command;
+
+import duke.Storage;
+import duke.TaskList;
+import duke.Ui;
+
+public class ExitCommand extends Command {
+
+    @Override
+    public void execute(TaskList tasks, Ui ui, Storage storage) {
+        return;
+    }
+
+    @Override
+    public boolean isExit() {
+        return true;
+    }
+
+}
