+++ src/main/java/duke/command/MessageCommand.java
@@ -1,3 +1,10 @@
+package duke.command;
+
+import duke.DukeException;
+import duke.Storage;
+import duke.TaskList;
+import duke.Ui;
+
 public class MessageCommand extends Command {
     private String message;
     public MessageCommand(String message) {
