+++ src/main/java/duke/task/Event.java
@@ -1,3 +1,7 @@
+package duke.task;
+
+import duke.task.Task;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
